Modify the types of primary key and add some attributes in MenuEntity, OrderEntity and UserAddressEntity. Add enums. Remove the dependency version of 'jaxb-api'.

+++ pizza-data/src/main/java/edu.ecnu.scsse.pizza.data/repository/MenuJpaRepository.java
@@ -3,6 +3,6 @@ package edu.ecnu.scsse.pizza.data.repository;
 import edu.ecnu.scsse.pizza.data.domain.MenuEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuJpaRepository extends JpaRepository<MenuEntity,Long> {
+public interface MenuJpaRepository extends JpaRepository<MenuEntity,Integer> {
 
 }
